fix: 타입 사이의 spacing 의 크기를 dp 로 적용

+++ android/app/src/main/java/poke/rogue/helper/presentation/dex/PokemonTypesAdapter.kt
@@ -24,7 +24,7 @@ class PokemonTypesAdapter(private val context: Context, private val viewGroup: V
                             LinearLayout.LayoutParams.WRAP_CONTENT,
                             1f,
                         ).apply {
-                            setMargins(horizontalMargin = config.spacingBetweenTypes.toInt())
+                            setMargins(horizontalMargin = config.spacingBetweenTypes)
                         }
                     TypeChip.setTypeUiConfiguration(
                         view = this,
